requirements.txt
accelerate
aioboto3
aiohttp
python-dotenv
python-multipart
fastapi
uvicorn[standard]
websockets
PyPDF2
python-docx
docx2txt
python-pptx
openpyxl
striprtf
odfpy
ebooklib
beautifulsoup4
# database packages
graspologic
openai
nano-vectordb
psycopg[binary,pool]
sqlalchemy
tenacity
networkx
python-jose
passlib
bcrypt
pyjwt
aiosqlite
matplotlib

# LLM packages
tiktoken
torch
transformers
xxhash
